Remoção de imagem do filesystem após exclusão em algum model

+++ administrador/src/alunos/views.py
@@ -138,11 +138,14 @@ def student_form_view(request, id=0):
 
 @login_required(login_url='login')
 def student_delete_view(request, id=0):
-	try:
-		student = Aluno.objects.get(id=id)
-		treino = student.cod_treino
-		response = requests.post('http://127.0.0.1:5000/api/delete', data={'faceID': treino})
-		if response.status_code == 200:
-			student.delete()
-	finally:
+	if request.method == 'POST':
+		try:
+			student = Aluno.objects.get(id=id)
+			treino = student.cod_treino
+			response = requests.post('http://127.0.0.1:5000/api/delete', data={'faceID': treino})
+			if response.status_code == 200:
+				student.delete()
+		finally:
+			return redirect('/alunos/')
+	else:
 		return redirect('/alunos/')
